requirements.txt
flask==2.3.3
opencv-python==4.8.0.74
numpy==1.24.3
ffmpeg-python==0.2.0
python-dotenv==1.0.0
Werkzeug==2.3.7 